requirements.txt
fastapi[standard]
SQLAlchemy
pydantic-settings
alembic
pyodbc